refactored a few changes
