--- api-review-khushi/src/main/java/therapyapp/Main.java
@@ -1,5 +1,0 @@
-package therapyapp;
-
-public class Main {
-    
-}
