--- api-review-khushi/src/main/java/therapyapp/TherapyAppStack.java
@@ -1,5 +1,0 @@
-package therapyapp;
-
-public class TherapyAppStack {
-    
-}
