--- api-review-khushi/src/main/java/therapyapp/handler/therapists/JournalAccessHandler.java
@@ -1,5 +1,0 @@
-package therapyapp.handler.therapists;
-
-public class JournalAccessHandler {
-    
-}
