--- api-review-khushi/src/main/java/therapyapp/model/JournalEntry.java
@@ -1,64 +1,0 @@
-package therapyapp.model;
-
-import java.util.Date;
-
-public class JournalEntry {
-    private String journalId;
-    private String userId;
-    private String content;
-    private Date createdAt;
-    private Date updatedAt;
-
-    // Constructors
-    public JournalEntry() {
-    }
-
-    public JournalEntry(String journalId, String userId, String content, Date createdAt, Date updatedAt) {
-        this.journalId = journalId;
-        this.userId = userId;
-        this.content = content;
-        this.createdAt = createdAt;
-        this.updatedAt = updatedAt;
-    }
-
-    // Getters and Setters
-    public String getJournalId() {
-        return journalId;
-    }
-
-    public void setJournalId(String journalId) {
-        this.journalId = journalId;
-    }
-
-    public String getUserId() {
-        return userId;
-    }
-
-    public void setUserId(String userId) {
-        this.userId = userId;
-    }
-
-    public String getContent() {
-        return content;
-    }
-
-    public void setContent(String content) {
-        this.content = content;
-    }
-
-    public Date getCreatedAt() {
-        return createdAt;
-    }
-
-    public void setCreatedAt(Date createdAt) {
-        this.createdAt = createdAt;
-    }
-
-    public Date getUpdatedAt() {
-        return updatedAt;
-    }
-
-    public void setUpdatedAt(Date updatedAt) {
-        this.updatedAt = updatedAt;
-    }
-}
