--- api-review-khushi/src/main/java/therapyapp/model/Message.java
@@ -1,64 +1,0 @@
-package therapyapp.model;
-
-import java.util.Date;
-
-public class Message {
-    private String messageId;
-    private String senderId;
-    private String receiverId;
-    private String content;
-    private Date sentAt;
-
-    // Constructors
-    public Message() {
-    }
-
-    public Message(String messageId, String senderId, String receiverId, String content, Date sentAt) {
-        this.messageId = messageId;
-        this.senderId = senderId;
-        this.receiverId = receiverId;
-        this.content = content;
-        this.sentAt = sentAt;
-    }
-
-    // Getters and Setters
-    public String getMessageId() {
-        return messageId;
-    }
-
-    public void setMessageId(String messageId) {
-        this.messageId = messageId;
-    }
-
-    public String getSenderId() {
-        return senderId;
-    }
-
-    public void setSenderId(String senderId) {
-        this.senderId = senderId;
-    }
-
-    public String getReceiverId() {
-        return receiverId;
-    }
-
-    public void setReceiverId(String receiverId) {
-        this.receiverId = receiverId;
-    }
-
-    public String getContent() {
-        return content;
-    }
-
-    public void setContent(String content) {
-        this.content = content;
-    }
-
-    public Date getSentAt() {
-        return sentAt;
-    }
-
-    public void setSentAt(Date sentAt) {
-        this.sentAt = sentAt;
-    }
-}
