--- api-review-khushi/src/main/java/therapyapp/model/User.java
@@ -1,52 +1,0 @@
-package therapyapp.model;
-
-public class User {
-    private String userId;
-    private String name;
-    private String email;
-    private String role;  // "client" or "therapist"
-
-    // Constructors
-    public User() {
-    }
-
-    public User(String userId, String name, String email, String role) {
-        this.userId = userId;
-        this.name = name;
-        this.email = email;
-        this.role = role;
-    }
-
-    // Getters and Setters
-    public String getUserId() {
-        return userId;
-    }
-
-    public void setUserId(String userId) {
-        this.userId = userId;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-    public String getRole() {
-        return role;
-    }
-
-    public void setRole(String role) {
-        this.role = role;
-    }
-}
