--- api-review-khushi/src/test/java/com/myorg/Test1Test.java
@@ -1,26 +1,0 @@
-// package com.myorg;
-
-// import software.amazon.awscdk.App;
-// import software.amazon.awscdk.assertions.Template;
-// import java.io.IOException;
-
-// import java.util.HashMap;
-
-// import org.junit.jupiter.api.Test;
-
-// example test. To run these tests, uncomment this file, along with the
-// example resource in java/src/main/java/com/myorg/Test1Stack.java
-// public class Test1Test {
-
-//     @Test
-//     public void testStack() throws IOException {
-//         App app = new App();
-//         Test1Stack stack = new Test1Stack(app, "test");
-
-//         Template template = Template.fromStack(stack);
-
-//         template.hasResourceProperties("AWS::SQS::Queue", new HashMap<String, Number>() {{
-//           put("VisibilityTimeout", 300);
-//         }});
-//     }
-// }
