+++ khushi-api-review/src/main/java/therapyapp/Main.java
@@ -1,0 +1,9 @@
+package therapyapp;
+
+
+public class Main {
+    public static void main(String[] args) {
+        // Placeholder entry point if needed for local testing
+        System.out.println("Therapy Journaling API running.");
+    }
+}
