+++ khushi-api-review/src/main/java/therapyapp/handler/auth/AuthHandler.java
