+++ khushi-api-review/src/main/java/therapyapp/handler/clients/JournalHandler.java
