+++ khushi-api-review/src/main/java/therapyapp/handler/sessions/SessionHandler.java
