+++ khushi-api-review/src/main/java/therapyapp/handler/therapists/JournalAccessHandler.java
@@ -1,0 +1,27 @@
+package therapyapp.handler.therapists;
+
+import java.util.HashMap;
+import java.util.Map;
+
+import com.amazonaws.services.lambda.runtime.Context;
+import com.amazonaws.services.lambda.runtime.RequestHandler;
+
+public class JournalAccessHandler implements RequestHandler<Map<String, Object>, Map<String, Object>> {
+
+    @Override
+    public Map<String, Object> handleRequest(Map<String, Object> input, Context context) {
+        context.getLogger().log("Processing journal access request");
+        Map<String, Object> response = new HashMap<>();
+
+        try {
+            String accessType = "PENDING";  // Placeholder for actual logic
+            response.put("statusCode", 200);
+            response.put("body", "Journal access status updated: " + accessType);
+        } catch (Exception e) {
+            response.put("statusCode", 500);
+            response.put("body", "Error processing journal access request");
+        }
+
+        return response;
+    }
+}
