+++ khushi-api-review/src/main/java/therapyapp/handler/therapists/MappingHandler.java
