+++ khushi-api-review/src/main/java/therapyapp/model/Message.java
@@ -1,0 +1,61 @@
+package therapyapp.model;
+
+import java.time.LocalDateTime;
+
+public class Message {
+    private String messageId;
+    private String senderId;
+    private String receiverId;
+    private String messageContent;
+    private LocalDateTime timestamp;
+
+    public Message() {}
+
+    public Message(String messageId, String senderId, String receiverId, String messageContent, LocalDateTime timestamp) {
+        this.messageId = messageId;
+        this.senderId = senderId;
+        this.receiverId = receiverId;
+        this.messageContent = messageContent;
+        this.timestamp = timestamp;
+    }
+
+    public String getMessageId() {
+        return messageId;
+    }
+
+    public void setMessageId(String messageId) {
+        this.messageId = messageId;
+    }
+
+    public String getSenderId() {
+        return senderId;
+    }
+
+    public void setSenderId(String senderId) {
+        this.senderId = senderId;
+    }
+
+    public String getReceiverId() {
+        return receiverId;
+    }
+
+    public void setReceiverId(String receiverId) {
+        this.receiverId = receiverId;
+    }
+
+    public String getMessageContent() {
+        return messageContent;
+    }
+
+    public void setMessageContent(String messageContent) {
+        this.messageContent = messageContent;
+    }
+
+    public LocalDateTime getTimestamp() {
+        return timestamp;
+    }
+
+    public void setTimestamp(LocalDateTime timestamp) {
+        this.timestamp = timestamp;
+    }
+}
