+++ khushi-api-review/src/main/java/therapyapp/model/User.java
@@ -1,0 +1,49 @@
+package therapyapp.model;
+
+public class User {
+    private String userId;
+    private String email;
+    private String password;
+    private String role;  // CLIENT or THERAPIST
+
+    public User() {}
+
+    public User(String userId, String email, String password, String role) {
+        this.userId = userId;
+        this.email = email;
+        this.password = password;
+        this.role = role;
+    }
+
+    public String getUserId() {
+        return userId;
+    }
+
+    public void setUserId(String userId) {
+        this.userId = userId;
+    }
+
+    public String getEmail() {
+        return email;
+    }
+
+    public void setEmail(String email) {
+        this.email = email;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+
+    public void setPassword(String password) {
+        this.password = password;
+    }
+
+    public String getRole() {
+        return role;
+    }
+
+    public void setRole(String role) {
+        this.role = role;
+    }
+}
